requirements.txt
requests==2.22.0
urllib3==1.25.3
certifi==2019.9.11
hexdump==3.3
asn1==2.2.0
colored==1.4.0
argcomplete==1.10.0
pyusb==1.0.2
